Crate syntastica_themes

syntastica-themes

A collection of themes for syntastica.

See the project overview for more information.

Features

runtime-c (enabled by default) — Use the standard tree-sitter C runtime. See syntastica’s WebAssembly support for more information.

runtime-c2rust — Use the pure Rust tree-sitter runtime. See syntastica’s WebAssembly support for more information.

docs — Enable this when building the docs

Modules

gruvbox

The gruvbox themes in this module were extracted from https://github.com/ellisonleao/gruvbox.nvim

one

The onedark themes in this module were extracted from https://github.com/navarasu/onedark.nvim

Functions

from_str

Try to get a theme given its path as a string.