Crate syntastica_themes

syntastica-themes

A collection of themes for syntastica.

See the project overview for more information.

Features

runtime-c (enabled by default) — Use the standard tree-sitter C runtime. See syntastica’s WebAssembly support for more information.

runtime-c2rust — Use the pure Rust tree-sitter runtime. See syntastica’s WebAssembly support for more information.

docs — Enable this when building the docs

Modules

abscs

The abscs theme collection in this module was extracted from https://github.com/Abstract-IDE/Abstract-cs using auto_extract.py.

aurora

The aurora theme collection in this module was extracted from https://github.com/ray-x/aurora using auto_extract.py.

blue_moon

The blue_moon theme collection in this module was extracted from https://github.com/kyazdani42/blue-moon using auto_extract.py.

boo

The boo theme collection in this module was extracted from https://github.com/rockerBOO/boo-colorscheme-nvim using auto_extract.py.

catppuccin

The catppuccin theme collection in this module was extracted from https://github.com/catppuccin/nvim using auto_extract.py.

darcula

The darcula theme collection in this module was extracted from https://github.com/doums/darcula using auto_extract.py.

dracula

The dracula theme collection in this module was extracted from https://github.com/dracula/vim using auto_extract.py.

everblush

The everblush theme collection in this module was extracted from https://github.com/Everblush/nvim using auto_extract.py.

everforest

The everforest theme collection in this module was extracted from https://github.com/sainnhe/everforest using auto_extract.py.

falcon

The falcon theme collection in this module was extracted from https://github.com/fenetikm/falcon using auto_extract.py.

github

The github theme collection in this module was extracted from https://github.com/project0n/github-nvim-theme using auto_extract.py.

gruvbox

The gruvbox themes in this module were extracted from https://github.com/ellisonleao/gruvbox.nvim

material

The material theme collection in this module was extracted from https://github.com/marko-cerovac/material.nvim using auto_extract.py.

melange

The melange theme collection in this module was extracted from https://github.com/savq/melange-nvim using auto_extract.py.

minimal

The minimal theme collection in this module was extracted from https://github.com/Yazeed1s/minimal.nvim using auto_extract.py.

monochrome

The monochrome theme collection in this module was extracted from https://github.com/kdheepak/monochrome.nvim using auto_extract.py.

monokai

The monokai theme collection in this module was extracted from https://github.com/tanvirtin/monokai.nvim using auto_extract.py.

moonfly

The moonfly theme collection in this module was extracted from https://github.com/bluz71/vim-moonfly-colors using auto_extract.py.

moonlight

The moonlight theme collection in this module was extracted from https://github.com/shaunsingh/moonlight.nvim using auto_extract.py.

neon

The neon theme collection in this module was extracted from https://github.com/rafamadriz/neon using auto_extract.py.

nightfly

The nightfly theme collection in this module was extracted from https://github.com/bluz71/vim-nightfly-colors using auto_extract.py.

nord

The nord theme collection in this module was extracted from https://github.com/shaunsingh/nord.nvim using auto_extract.py.

oceanicnext

The oceanicnext theme collection in this module was extracted from https://github.com/mhartington/oceanic-next using auto_extract.py.

omni

The omni theme collection in this module was extracted from https://github.com/yonlu/omni using auto_extract.py.

one

The onedark themes in this module were extracted from https://github.com/navarasu/onedark.nvim

oxocarbon

The oxocarbon theme collection in this module was extracted from https://github.com/nyoom-engineering/oxocarbon.nvim using auto_extract.py.

solarized

The solarized theme collection in this module was extracted from https://github.com/overcache/NeoSolarized using auto_extract.py.

tokyo

The tokyo theme collection in this module was extracted from https://github.com/folke/tokyonight.nvim using auto_extract.py.

vscode

The vscode theme collection in this module was extracted from https://github.com/Mofiqul/vscode.nvim using auto_extract.py.

zephyr

The zephyr theme collection in this module was extracted from https://github.com/glepnir/zephyr-nvim using auto_extract.py.

Constants

THEMES

A list of all theme names as they are accepted by from_str.

Functions

from_str

Try to get a theme given its path as a string.